// ui/binmerchant/src/BinExport2.d.ts
export interface IBinExport2 {

    /** BinExport2 metaInformation */
    metaInformation?: (BinExport2.IMeta|null);

    /** BinExport2 expression */
    expression?: (BinExport2.IExpression[]|null);

    /** BinExport2 operand */
    operand?: (BinExport2.IOperand[]|null);

    /** BinExport2 mnemonic */
    mnemonic?: (BinExport2.IMnemonic[]|null);

    /** BinExport2 instruction */
    instruction?: (BinExport2.IInstruction[]|null);

    /** BinExport2 basicBlock */
    basicBlock?: (BinExport2.IBasicBlock[]|null);

    /** BinExport2 flowGraph */
    flowGraph?: (BinExport2.IFlowGraph[]|null);

    /** BinExport2 callGraph */
    callGraph?: (BinExport2.ICallGraph|null);

    /** BinExport2 stringTable */
    stringTable?: (string[]|null);

    /** BinExport2 addressComment */
    addressComment?: (BinExport2.IReference[]|null);

    /** BinExport2 comment */
    comment?: (BinExport2.IComment[]|null);

    /** BinExport2 stringReference */
    stringReference?: (BinExport2.IReference[]|null);

    /** BinExport2 expressionSubstitution */
    expressionSubstitution?: (BinExport2.IReference[]|null);

    /** BinExport2 section */
    section?: (BinExport2.ISection[]|null);

    /** BinExport2 library */
    library?: (BinExport2.ILibrary[]|null);

    /** BinExport2 dataReference */
    dataReference?: (BinExport2.IDataReference[]|null);

    /** BinExport2 module */
    module?: (BinExport2.IModule[]|null);
}

/** Represents a BinExport2. */
export class BinExport2 implements IBinExport2 {

    /**
     * Constructs a new BinExport2.
     * @param [properties] Properties to set
     */
    constructor(properties?: IBinExport2);

    /** BinExport2 metaInformation. */
    public metaInformation?: (BinExport2.IMeta|null);

    /** BinExport2 expression. */
    public expression: BinExport2.IExpression[];

    /** BinExport2 operand. */
    public operand: BinExport2.IOperand[];

    /** BinExport2 mnemonic. */
    public mnemonic: BinExport2.IMnemonic[];

    /** BinExport2 instruction. */
    public instruction: BinExport2.IInstruction[];

    /** BinExport2 basicBlock. */
    public basicBlock: BinExport2.IBasicBlock[];

    /** BinExport2 flowGraph. */
    public flowGraph: BinExport2.IFlowGraph[];

    /** BinExport2 callGraph. */
    public callGraph?: (BinExport2.ICallGraph|null);

    /** BinExport2 stringTable. */
    public stringTable: string[];

    /** BinExport2 addressComment. */
    public addressComment: BinExport2.IReference[];

    /** BinExport2 comment. */
    public comment: BinExport2.IComment[];

    /** BinExport2 stringReference. */
    public stringReference: BinExport2.IReference[];

    /** BinExport2 expressionSubstitution. */
    public expressionSubstitution: BinExport2.IReference[];

    /** BinExport2 section. */
    public section: BinExport2.ISection[];

    /** BinExport2 library. */
    public library: BinExport2.ILibrary[];

    /** BinExport2 dataReference. */
    public dataReference: BinExport2.IDataReference[];

    /** BinExport2 module. */
    public module: BinExport2.IModule[];

    /**
     * Creates a new BinExport2 instance using the specified properties.
     * @param [properties] Properties to set
     * @returns BinExport2 instance
     */
    public static create(properties?: IBinExport2): BinExport2;

    /**
     * Encodes the specified BinExport2 message. Does not implicitly {@link BinExport2.verify|verify} messages.
     * @param message BinExport2 message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: IBinExport2, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified BinExport2 message, length delimited. Does not implicitly {@link BinExport2.verify|verify} messages.
     * @param message BinExport2 message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: IBinExport2, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a BinExport2 message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns BinExport2
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): BinExport2;

    /**
     * Decodes a BinExport2 message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns BinExport2
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): BinExport2;

    /**
     * Verifies a BinExport2 message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a BinExport2 message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns BinExport2
     */
    public static fromObject(object: { [k: string]: any }): BinExport2;

    /**
     * Creates a plain object from a BinExport2 message. Also converts values to other types if specified.
     * @param message BinExport2
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: BinExport2, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this BinExport2 to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };

    /**
     * Gets the default type url for BinExport2
     * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
     * @returns The default type url
     */
    public static getTypeUrl(typeUrlPrefix?: string): string;
}

export namespace BinExport2 {

    /** Properties of a Meta. */
    interface IMeta {

        /** Meta executableName */
        executableName?: (string|null);

        /** Meta executableId */
        executableId?: (string|null);

        /** Meta architectureName */
        architectureName?: (string|null);

        /** Meta timestamp */
        timestamp?: (number|Long|null);
    }
// ...
    interface ICallGraph {

        /** CallGraph vertex */
        vertex?: (BinExport2.CallGraph.IVertex[]|null);

        /** CallGraph edge */
        edge?: (BinExport2.CallGraph.IEdge[]|null);
    }
// ...
    namespace CallGraph {

        /** Properties of a Vertex. */
        interface IVertex {

            /** Vertex address */
            address?: (number|Long|null);

            /** Vertex type */
            type?: (BinExport2.CallGraph.Vertex.Type|null);

            /** Vertex mangledName */
            mangledName?: (string|null);

            /** Vertex demangledName */
            demangledName?: (string|null);

            /** Vertex libraryIndex */
            libraryIndex?: (number|null);

            /** Vertex moduleIndex */
            moduleIndex?: (number|null);
        }
// ...
        namespace Vertex {

            /** Type enum. */
            enum Type {
                NORMAL = 0,
                LIBRARY = 1,
                IMPORTED = 2,
                THUNK = 3,
                INVALID = 4
            }
        }

        /** Properties of an Edge. */
        interface IEdge {

            /** Edge sourceVertexIndex */
            sourceVertexIndex?: (number|null);

            /** Edge targetVertexIndex */
            targetVertexIndex?: (number|null);
        }
// ...
    }

    /** Properties of an Expression. */
    interface IExpression {

        /** Expression type */
        type?: (BinExport2.Expression.Type|null);

        /** Expression symbol */
        symbol?: (string|null);

        /** Expression immediate */
        immediate?: (number|Long|null);

        /** Expression parentIndex */
        parentIndex?: (number|null);

        /** Expression isRelocation */
        isRelocation?: (boolean|null);
    }
// ...
    interface IOperand {

        /** Operand expressionIndex */
        expressionIndex?: (number[]|null);
    }
// ...
    interface IMnemonic {

        /** Mnemonic name */
        name?: (string|null);
    }
// ...
    interface IInstruction {

        /** Instruction address */
        address?: (number|Long|null);

        /** Instruction callTarget */
        callTarget?: ((number|Long)[]|null);

        /** Instruction mnemonicIndex */
        mnemonicIndex?: (number|null);

        /** Instruction operandIndex */
        operandIndex?: (number[]|null);

        /** Instruction rawBytes */
        rawBytes?: (Uint8Array|null);

        /** Instruction commentIndex */
        commentIndex?: (number[]|null);
    }
// ...
    interface IBasicBlock {

        /** BasicBlock instructionIndex */
        instructionIndex?: (BinExport2.BasicBlock.IIndexRange[]|null);
    }
// ...
    namespace BasicBlock {

        /** Properties of an IndexRange. */
        interface IIndexRange {

            /** IndexRange beginIndex */
            beginIndex?: (number|null);

            /** IndexRange endIndex */
            endIndex?: (number|null);
        }
// ...
    }

    /** Properties of a FlowGraph. */
    interface IFlowGraph {

        /** FlowGraph basicBlockIndex */
        basicBlockIndex?: (number[]|null);

        /** FlowGraph entryBasicBlockIndex */
        entryBasicBlockIndex?: (number|null);

        /** FlowGraph edge */
        edge?: (BinExport2.FlowGraph.IEdge[]|null);
    }
// ...
    interface IReference {

        /** Reference instructionIndex */
        instructionIndex?: (number|null);

        /** Reference instructionOperandIndex */
        instructionOperandIndex?: (number|null);

        /** Reference operandExpressionIndex */
        operandExpressionIndex?: (number|null);

        /** Reference stringTableIndex */
        stringTableIndex?: (number|null);
    }
// ...
    interface IDataReference {

        /** DataReference instructionIndex */
        instructionIndex?: (number|null);

        /** DataReference address */
        address?: (number|Long|null);
    }
// ...
    interface IComment {

        /** Comment instructionIndex */
        instructionIndex?: (number|null);

        /** Comment instructionOperandIndex */
        instructionOperandIndex?: (number|null);

        /** Comment operandExpressionIndex */
        operandExpressionIndex?: (number|null);

        /** Comment stringTableIndex */
        stringTableIndex?: (number|null);

        /** Comment repeatable */
        repeatable?: (boolean|null);

        /** Comment type */
        type?: (BinExport2.Comment.Type|null);
    }
// ...
    interface ISection {

        /** Section address */
        address?: (number|Long|null);

        /** Section size */
        size?: (number|Long|null);

        /** Section flagR */
        flagR?: (boolean|null);

        /** Section flagW */
        flagW?: (boolean|null);

        /** Section flagX */
        flagX?: (boolean|null);
    }
// ...
    interface ILibrary {

        /** Library isStatic */
        isStatic?: (boolean|null);

        /** Library loadAddress */
        loadAddress?: (number|Long|null);

        /** Library name */
        name?: (string|null);
    }
// ...
    interface IModule {

        /** Module name */
        name?: (string|null);
    }
// ...
}